DJANGO/DjangoRestFramework/introrest/requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
Unidecode==1.1.1
wrapt==1.11.2
